+++ NEWS.md
@@ -1,6 +1,8 @@
 # mirai 1.0.0.9010 (development)
 
 * Adds `x[]` as a more efficient equivalent of `call_mirai_(x)$data`.
+  + Use of `call_mirai_()` is consequently deprecated.
+* Adds `collect_mirai()` as a more efficient equivalent of `call_mirai(x)$data`.
 * Adds `mirai_map()` for asynchronous parallel/distributed map using mirai. Allows recovery from partial failure or else early stopping, together with optional progress reporting.
   + `x[]` collects the results of a mirai_map `x`, waiting for all asynchronous operations to complete.
   + `x[.progress]` collects the results whilst showing a text progress bar.